fix bug and optimized code
